requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
RPi.GPIO==0.7.1 